gitpulse/requirements.txt
textual>=0.50.0
rich>=13.0.0
gitpython>=3.1.0
